Re-organized utilities
