+++ src/features/common/Footer/index.tsx
@@ -3,7 +3,7 @@ import AppleIcon from '../../../assets/images/footer/AppStore';
 import GooglePlayIcon from '../../../assets/images/footer/GooglePlay';
 import UNEPLogo from '../../../assets/images/footer/UNEPLogo';
 import World from '../../../assets/images/footer/World';
-import getLanguageName from '../../../utils/getLanguageName';
+import getLanguageName from '../../../utils/language/getLanguageName';
 import styles from './Footer.module.scss';
 import SelectLanguageAndCountry from './SelectLanguageAndCountry';
 import i18next from '../../../../i18n';
