+++ src/features/public/Donations/components/PopupProject.tsx
@@ -2,14 +2,14 @@ import Modal from '@material-ui/core/Modal';
 import { Elements } from '@stripe/react-stripe-js';
 import React, { ReactElement, Ref } from 'react';
 import Sugar from 'sugar';
-import { getCountryDataBy } from '../../../../utils/countryUtils';
+import { getCountryDataBy } from '../../../../utils/countryCurrency/countryUtils';
 import getImageUrl from '../../../../utils/getImageURL';
-import getStripe from '../../../../utils/getStripe';
-import { ThemeContext } from '../../../../utils/themeContext';
+import getStripe from '../../../../utils/stripe/getStripe';
+import { ThemeContext } from '../../../../theme/themeContext';
 import DonationsPopup from './../screens/DonationsPopup';
 import styles from './../styles/Projects.module.scss';
 import i18next from '../../../../../i18n';
-import getFormatedCurrency from '../../../../utils/getFormattedCurrency';
+import getFormatedCurrency from '../../../../utils/countryCurrency/getFormattedCurrency';
 
 const { useTranslation } = i18next;
 interface Props {
