+++ src/features/public/Donations/components/SingleProjectDetails.tsx
@@ -12,15 +12,15 @@ import BlackTree from '../../../../assets/images/icons/project/BlackTree';
 import Email from '../../../../assets/images/icons/project/Email';
 import Location from '../../../../assets/images/icons/project/Location';
 import WorldWeb from '../../../../assets/images/icons/project/WorldWeb';
-import { getCountryDataBy } from '../../../../utils/countryUtils';
+import { getCountryDataBy } from '../../../../utils/countryCurrency/countryUtils';
 import getImageUrl from '../../../../utils/getImageURL';
-import getStripe from '../../../../utils/getStripe';
-import { ThemeContext } from '../../../../utils/themeContext';
+import getStripe from '../../../../utils/stripe/getStripe';
+import { ThemeContext } from '../../../../theme/themeContext';
 import ProjectContactDetails from '../components/projectDetails/ProjectContactDetails';
 import DonationsPopup from '../screens/DonationsPopup';
 import styles from './../styles/ProjectDetails.module.scss';
 import i18next from '../../../../../i18n';
-import getFormatedCurrency from '../../../../utils/getFormattedCurrency';
+import getFormatedCurrency from '../../../../utils/countryCurrency/getFormattedCurrency';
 
 const { useTranslation } = i18next;
 interface Props {
