+++ src/features/public/Donations/components/treeDonation/SelectTaxDeductionCountryModal.tsx
@@ -5,8 +5,8 @@ import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Modal from '@material-ui/core/Modal';
 import RadioGroup from '@material-ui/core/RadioGroup';
 import React, { useState } from 'react';
-import { getCountryDataBy } from '../../../../../utils/countryUtils';
-import { ThemeContext } from '../../../../../utils/themeContext';
+import { getCountryDataBy } from '../../../../../utils/countryCurrency/countryUtils';
+import { ThemeContext } from '../../../../../theme/themeContext';
 import GreenRadio from '../../../../common/InputTypes/GreenRadio';
 import i18next from '../../../../../../i18n';
 let styles = require('./../../styles/SelectCurrencyModal.module.scss');
