+++ src/utils/apiRequests/getSessionId.ts
