+++ src/utils/countryCurrency/countriesData.json
