+++ src/utils/countryCurrency/getFormattedCurrency.ts
